Implementacao de Handler

+++ SpringProject/src/main/java/mjv/controllers/AlunoController.java
@@ -1,8 +1,6 @@
 package mjv.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -13,11 +11,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-
 import mjv.entity.Aluno;
 import mjv.entity.enums.NivelAluno;
+import mjv.exceptions.NotFoundException;
+import mjv.exceptions.PostException;
 import mjv.repositories.AlunoRepository;
-
 
 @RestController // @Controller + @RespondeBody
 @RequestMapping("/mjv/aluno")
@@ -27,6 +25,10 @@ public class AlunoController {
 	
 	@PostMapping("/post")
 	public void create(@RequestParam("id") Integer id, @RequestParam("nivel") NivelAluno nivel, @RequestBody Aluno aluno) {
+		Aluno alunoDb = findById(id);
+		if(alunoDb != null) {
+			throw new PostException(id);
+		}
 		aluno.setId(id);
 		aluno.setNivel(nivel);
 		aluno.getPessoa().setId(id);
@@ -35,6 +37,10 @@ public class AlunoController {
 	
 	@PutMapping("/put")
 	public void update(@RequestParam("id") Integer id, @RequestParam("nivel") NivelAluno nivel, @RequestBody Aluno aluno) {
+		Aluno alunoDb = findById(id);
+		if(alunoDb == null) {
+			throw new NotFoundException(id, "PUT");
+		}
 		aluno.setId(id);
 		aluno.setNivel(nivel);
 		aluno.getPessoa().setId(id);
@@ -43,6 +49,10 @@ public class AlunoController {
 	
 	@PatchMapping("/patch")
 	public void patch(@RequestParam("id") Integer id, @RequestBody Aluno aluno) {
+		Aluno alunoDb = findById(id);
+		if(alunoDb == null) {
+			throw new NotFoundException(id, "PATCH");
+		}
 		aluno.setId(id);
 		aluno.getPessoa().setId(id);
 		Aluno alunoVelho = findById(id);
@@ -51,13 +61,12 @@ public class AlunoController {
 	}
 	
 	@DeleteMapping("/delete")
-	public ResponseEntity delete(@RequestParam("id") Integer id) {
-		try {
-			repo.deleteById(id);
-			return new ResponseEntity(HttpStatus.OK);
-		} catch(Exception ex) {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ex.getMessage());
+	public void delete(@RequestParam("id") Integer id) {
+		Aluno aluno = findById(id);
+		if(aluno == null) {
+			throw new NotFoundException(id, "DELETE");
 		}
+		repo.deleteById(id);
 	}
 	
 	@GetMapping("/getById")
@@ -86,10 +95,20 @@ public class AlunoController {
 		if(aluno.getContato1().getParentesco()	 != null) alunoVelho.getContato1().setParentesco(aluno.getContato1().getParentesco());
 		if(aluno.getContato1().getTelefone()	 != null) alunoVelho.getContato1().setTelefone(aluno.getContato1().getTelefone());
 		if(aluno.getContato1().getEmail() 	 	 != null) alunoVelho.getContato1().setEmail(aluno.getContato1().getEmail());
-		if(aluno.getContato2().getNome()	 	 != null) alunoVelho.getContato2().setNome(aluno.getContato2().getNome());
-		if(aluno.getContato2().getParentesco()	 != null) alunoVelho.getContato2().setParentesco(aluno.getContato2().getParentesco());
-		if(aluno.getContato2().getTelefone()	 != null) alunoVelho.getContato2().setTelefone(aluno.getContato2().getTelefone());
-		if(aluno.getContato2().getEmail() 	 	 != null) alunoVelho.getContato2().setEmail(aluno.getContato2().getEmail());
+		alunoVelho = updaterContato2(aluno, alunoVelho);
 		return alunoVelho;
+	}
+	
+	private Aluno updaterContato2(Aluno aluno, Aluno alunoVelho) {
+		if(aluno.getContato2() != null && alunoVelho.getContato2() == null) {
+			alunoVelho.setContato2(aluno.getContato2());
+			return alunoVelho;
+		} else {
+			if(aluno.getContato2().getNome()	 	 != null) alunoVelho.getContato2().setNome(aluno.getContato2().getNome());
+			if(aluno.getContato2().getParentesco()	 != null) alunoVelho.getContato2().setParentesco(aluno.getContato2().getParentesco());
+			if(aluno.getContato2().getTelefone()	 != null) alunoVelho.getContato2().setTelefone(aluno.getContato2().getTelefone());
+			if(aluno.getContato2().getEmail() 	 	 != null) alunoVelho.getContato2().setEmail(aluno.getContato2().getEmail());
+			return alunoVelho;
+		}
 	}
 }
